Use uint32 record offsets and improve overflow checks

+++ cpp/include/seqwin/filter.hpp
@@ -12,7 +12,7 @@ void get_penalty(
     const Kmer* kmers,
     Node* nodes,
     std::size_t n_nodes,
-    const std::size_t* record_offsets,
+    const std::uint32_t* record_offsets,
     std::size_t n_record_offsets,
     const bool* is_targets,
     std::size_t n_assemblies,
